Split endoflife.date and releasedata models in separate files
This makes the code easier to read.
