+++ src/github-releases.py
@@ -3,7 +3,7 @@ import logging
 import subprocess
 import sys
 
-from common import dates, endoflife
+from common import dates, endoflife, releasedata
 
 """Fetches versions from GitHub releases using the GraphQL API and the GitHub CLI.
 
@@ -44,7 +44,7 @@ query($endCursor: String) {
 
 p_filter = sys.argv[1] if len(sys.argv) > 1 else None
 for product_name in endoflife.list_products(METHOD, p_filter):
-    product = endoflife.Product(product_name)
+    product = releasedata.Product(product_name)
     product_frontmatter = endoflife.ProductFrontmatter(product.name)
     for config in product_frontmatter.get_auto_configs(METHOD):
         for page in fetch_releases(config.url):
