+++ src/plesk.py
@@ -1,12 +1,12 @@
 from bs4 import BeautifulSoup
-from common import dates, endoflife, http
+from common import dates, http, releasedata
 
 """Fetches versions from Plesk's change log.
 
 Only 18.0.20.3 and later will be picked up, as the format of the change log for 18.0.20 and 18.0.19 are different and
 there is no entry for GA of version 18.0.18 and older."""
 
-product = endoflife.Product("plesk")
+product = releasedata.Product("plesk")
 response = http.fetch_url("https://docs.plesk.com/release-notes/obsidian/change-log")
 soup = BeautifulSoup(response.text, features="html5lib")
